Add game board render on the web page.
This commit adds game board on the web page as html divs. Each div is one of the TicTacToe fields, which could be occupied by a "X" or "O" sign. This commit also stylize the web page a bit. Moreover js file can render actual game state on the web page.